used toggle for index page and removed eth nyc stuff. Also moved toggle
